% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{Calculate the Steady-State Joint Distribution for a Population}
\usage{
calc_dist(
  incid = 0.1,
  vac = list(p_comm = 0.5, p_org = 0.5, eff = 0.7),
  inf = list(p_symp = 0.5, t_symp = 10, t_presymp = 3),
  test = list(p_symp = 0.95, p_asymp = 1/7),
  detect = list(sens = 0.85, spec = 1)
)
}
\arguments{
\item{incid}{`[numeric(1)]` Incidence rate per 100k per day in the community}

\item{vac}{`[list(3)]` A named list containing vaccination parameters:
\describe{
  \item{p_comm `[numeric(1)]`}{Proportion vaccinated in the community}
  \item{p_org `[numeric(1)]`}{Proportion vaccinated in the organization of interest}
  \item{eff `[numeric(1)]`}{Vaccine efficacy}
}}

\item{inf}{`[list(3)]` A named list containing infection parameters:
\describe{
  \item{p_symp `[numeric(1)]`}{Proportion of infections eventually showing symptoms}
  \item{t_inf `[numeric(1)]`}{Duration of infectious period}
  \item{t_presymp `[numeric(1)]`}{Duration of presymptomatic period}
}}

\item{test}{`[list(2)]` A named list containing testing parameters:
\describe{
  \item{p_symp `[numeric(1)]`}{Probability of being tested if symptomatic}
  \item{p_asymp `[numeric(1)]`}{Probability of being tested if asymptomatic}
}}

\item{detect}{`[list(2)]` A named list containing detection parameters:
\describe{
  \item{sens `[numeric(1)]`}{Test sensitivity}
  \item{spec `[numeric(1)]`}{Test specificity}
}}
}
\description{
`calc_dist()` calculates the discrete joint distribution of vaccination,
infection, symptoms, tests, and detections in a population.
}
