% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist_cache}
\alias{calc_dist_cache}
\title{Cached `calc_dist()`}
\usage{
calc_dist_cache(params)
}
\arguments{
\item{params}{A named `list` of arguments to `calc_dist()`}
}
\value{
A `reactive` expression
}
\description{
A version of `calc_dist()` that caches outputs based on previous inputs. This
is done using `bindCache()`, which allows the cache to be shared across user
sessions. It returns a reactive expression.
}
\keyword{internal}
