% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidscreen-package.R
\docType{package}
\name{covidscreen-package}
\alias{covidscreen}
\alias{covidscreen-package}
\title{covidscreen: Simulate Asymptomatic Screening for COVID-19}
\description{
The {covidscreen} package provides simple simulations for understanding the effects of assurance testing/screening on a population. It aims to help decision-makers determine whether and how to test for COVID-19 with limited resources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jesse-smith/covidscreen}
  \item Report bugs at \url{https://github.com/jesse-smith/covidscreen/issues}
}

}
\author{
\strong{Maintainer}: Jesse Smith \email{Jesse.Smith@stjude.org}

Authors:
\itemize{
  \item Tomi Mori \email{Tomi.Mori@stjude.org}
}

}
\keyword{internal}
